fix to move config.go for using in api side
