+++ cmd/web/handlers.go
@@ -3,6 +3,7 @@ package main
 import (
 	"github.com/go-chi/chi/v5"
 	"myapp/internal/cards"
+	config2 "myapp/internal/config"
 	"myapp/internal/models"
 	"net/http"
 	"strconv"
@@ -103,7 +104,7 @@ func (app *application) PaymentSucceeded(w http.ResponseWriter, r *http.Request)
 		return
 	}
 
-	config := LoadConfig()
+	config := config2.LoadConfig()
 	// create new transaction
 	txn := models.Transaction{
 		Amount:              txnData.PaymentAmount,
@@ -151,7 +152,7 @@ func (app *application) VirtualTerminalPaymentSucceeded(w http.ResponseWriter, r
 		return
 	}
 
-	config := LoadConfig()
+	config := config2.LoadConfig()
 	// create new transaction
 	txn := models.Transaction{
 		Amount:              txnData.PaymentAmount,
@@ -260,5 +261,11 @@ func (app *application) BronzePlan(w http.ResponseWriter, r *http.Request) {
 	}); err != nil {
 		app.errorLog.Println(err)
 	}
+}
 
+func (app *application) BronzePlanReceipt(w http.ResponseWriter, r *http.Request) {
+
+	if err := app.renderTemplate(w, r, "receipt-plan", &templateData{}); err != nil {
+		app.errorLog.Println(err)
+	}
 }
